# apps/core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Search, fetch, render and speak for every product, plus the admin API of the applications registry.",
        "title": "Tornade",
        "contact": {},
        "version": "0.3.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/admin/apps": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Every registered application, key withheld",
                "operationId": "listApps",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/registry.AppListDTO"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Admit an application and hand its key out once",
                "operationId": "registerApp",
                "parameters": [
                    {
                        "description": "The issuer name the application signs as",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/registry.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/registry.CredentialsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/registry.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/admin/apps/{name}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin"
                ],
                "summary": "End an application's access at once",
                "operationId": "revokeApp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/registry.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/admin/apps/{name}/rotate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Mint an application a new key; the old one works for a day",
                "operationId": "rotateAppKey",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/registry.CredentialsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/registry.ErrorDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/registry.ErrorDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "registry.AppDTO": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "grace_until": {
                    "type": "string"
                },
                "last4": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "revoked_at": {
                    "type": "string"
                },
                "rotated_at": {
                    "type": "string"
                }
            }
        },
        "registry.AppListDTO": {
            "type": "object",
            "properties": {
                "apps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/registry.AppDTO"
                    }
                }
            }
        },
        "registry.CredentialsDTO": {
            "type": "object",
            "properties": {
                "grace_until": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "registry.ErrorDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "registry.RegisterRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}